+++ packages/embed/e2e/spec/functional/callbacks.spec.ts
@@ -1,0 +1,10 @@
+describe('Callbacks', () => {
+  it('should alert onReady on form load', () => {
+    const stub = cy.stub()
+    cy.on('window:alert', stub)
+
+    cy.visit('/callbacks.html').then(() => {
+      expect(stub).to.be.calledWith('onReady')
+    })
+  })
+})