+++ packages/embed/e2e/spec/functional/widget.spec.ts
@@ -7,4 +7,10 @@ describe('Widget', () => {
     cy.get('.typeform-widget iframe').should('be.visible')
     cy.get('.typeform-widget iframe').invoke('attr', 'src').should('contain', 'form.typeform.com/to/')
   })
+
+  it('should pass options as query param', () => {
+    cy.get('.typeform-widget iframe')
+      .invoke('attr', 'src')
+      .should('contain', 'typeform-embed=embed-widget&typeform-source=localhost&typeform-medium=unit-test')
+  })
 })