+++ packages/embed/src/initializers/build-options-from-attributes.ts
@@ -1,0 +1,16 @@
+import { loadOptionsFromAttributes } from '../utils/load-options-from-attributes'
+
+export const buildOptionsFromAttributes = (element: HTMLElement) => {
+  return loadOptionsFromAttributes(element, {
+    source: 'string',
+    medium: 'string',
+    mediumVersion: 'string',
+    hideFooter: 'boolean',
+    hideHeaders: 'boolean',
+    opacity: 'integer',
+    disableTracking: 'boolean',
+    onReady: 'function',
+    onSubmit: 'function',
+    onQuestionChanged: 'function',
+  })
+}