+++ packages/embed/src/utils/build-iframe-src.ts
@@ -1,5 +1,54 @@
 import { EmbedType, UrlOptions } from '../base'
 
-export const buildIframeSrc = (formId: string, _type: EmbedType, _options: UrlOptions) => {
-  return `https://form.typeform.com/to/${formId}`
+import { removeUndefinedKeys } from './remove-undefined-keys'
+
+const defaultUrlOptions: UrlOptions = {
+  source: window?.location?.hostname.replace(/^www\./, ''),
+  medium: 'embed-sdk',
+}
+
+const addDefaultUrlOptions = (options: UrlOptions): UrlOptions => {
+  return { ...defaultUrlOptions, ...removeUndefinedKeys(options) }
+}
+
+const typesToEmbed: Record<EmbedType, string> = {
+  widget: 'embed-widget', // TODO: when widget is full page use 'embed-fullpage'
+  popup: 'popup-blank',
+  slider: 'popup-drawer',
+  popover: 'popup-popover',
+  'side-tab': 'popup-side-panel',
+}
+
+const mapOptionsToQueryParams = (type: EmbedType, embedId: string, options: UrlOptions): Record<string, any> => ({
+  'typeform-embed-id': embedId,
+  'typeform-embed': typesToEmbed[type],
+  'typeform-source': options.source,
+  'typeform-medium': options.medium,
+  'typeform-medium-version': options.mediumVersion,
+  'embed-hide-footer': options.hideFooter ? 'true' : undefined,
+  'embed-hide-headers': options.hideHeaders ? 'true' : undefined,
+  'embed-opacity': options.opacity,
+  'disable-tracking': options.disableTracking ? 'true' : undefined,
+})
+
+export const buildIframeSrc = (params: BuildIframeSrcOptions): string => {
+  const { formId, type, embedId, options } = params
+
+  const queryParams = mapOptionsToQueryParams(type, embedId, addDefaultUrlOptions(options))
+
+  const url = new URL(`https://form.typeform.com/to/${formId}`)
+  Object.entries(queryParams)
+    .filter(([, paramValue]) => paramValue != null)
+    .forEach(([paramName, paramValue]) => {
+      url.searchParams.set(paramName, paramValue)
+    })
+
+  return url.href
+}
+
+type BuildIframeSrcOptions = {
+  formId: string
+  embedId: string
+  type: EmbedType
+  options: UrlOptions
 }