+++ packages/embed/src/utils/create-iframe/create-iframe.spec.ts
@@ -4,22 +4,31 @@ import { EmbedType } from '../../base'
 
 import { createIframe } from './create-iframe'
 
+jest.mock('./generate-embed-id', () => ({ generateEmbedId: () => 'random-id' }))
+
 describe('create-iframe', () => {
   describe('#createIframe', () => {
-    let iframe: HTMLIFrameElement
+    let iframe = null as any
+
     const buildIframeSrcMock = jest
       .spyOn(require('./../build-iframe-src'), 'buildIframeSrc')
       .mockImplementation(() => 'http://iframe-src/')
     const createElementMock = jest.spyOn(document, 'createElement')
     const triggerIframeRedrawMock = jest.spyOn(require('./trigger-iframe-redraw'), 'triggerIframeRedraw')
-    const options = {}
+    const options = { onReady: jest.fn(), onSubmit: jest.fn(), onQuestionChanged: jest.fn() }
 
     beforeEach(() => {
       iframe = createIframe('form-id', EmbedType.Widget, options)
     })
 
     it('should call buildIframeSrc', () => {
-      expect(buildIframeSrcMock).toHaveBeenCalledWith('form-id', 'widget', options)
+      expect(buildIframeSrcMock).toHaveBeenCalledTimes(1)
+      expect(buildIframeSrcMock).toHaveBeenCalledWith({
+        embedId: 'random-id',
+        formId: 'form-id',
+        options,
+        type: 'widget',
+      })
     })
 
     it('should create new iframe element', () => {
@@ -36,6 +45,27 @@ describe('create-iframe', () => {
       fireEvent(iframe, new Event('load'))
 
       expect(triggerIframeRedrawMock).toHaveBeenCalled()
+    })
+
+    it('should call form-ready handler', async () => {
+      window.postMessage({ type: 'form-ready', embedId: 'random-id' }, '*')
+      await new Promise((resolve) => setTimeout(resolve))
+
+      expect(options.onReady).toBeCalled()
+    })
+
+    it('should call form-screen-changed handler', async () => {
+      window.postMessage({ type: 'form-screen-changed', embedId: 'random-id' }, '*')
+      await new Promise((resolve) => setTimeout(resolve))
+
+      expect(options.onQuestionChanged).toBeCalled()
+    })
+
+    it('should call form-submit handler', async () => {
+      window.postMessage({ type: 'form-submit', response_id: 'test-response-id', embedId: 'random-id' }, '*')
+      await new Promise((resolve) => setTimeout(resolve))
+
+      expect(options.onSubmit).toBeCalledWith({ responseId: 'test-response-id' })
     })
   })
 })