+++ packages/embed/src/utils/remove-undefined-keys.ts
@@ -1,0 +1,5 @@
+export const removeUndefinedKeys = (obj: Record<string, any>): Record<string, any> => {
+  return Object.entries(obj)
+    .filter(([, value]) => value != null)
+    .reduce((acc, [key, value]) => ({ ...acc, [key]: value }), {})
+}