+++ packages/embed/CHANGELOG.md
@@ -1,3 +1,37 @@
+# [@typeform/embed-v1.0.0-next.2](https://github.com/Typeform/embed/compare/@typeform/embed-v1.0.0-next.1...@typeform/embed-v1.0.0-next.2) (2021-02-22)
+
+
+### Bug Fixes
+
+* **DIST-730:** Fix redraw issue ([a65076d](https://github.com/Typeform/embed/commit/a65076d057e003f1664f701633a43393d3b270cc))
+
+
+### Features
+
+* **DIST-0:** Added config for cc velocity ([#170](https://github.com/Typeform/embed/issues/170)) ([02f7e71](https://github.com/Typeform/embed/commit/02f7e710b6184fe625f2f909a5abead9681acfb6))
+* **DIST-701:** Add functional tests and update demo ([e3edc97](https://github.com/Typeform/embed/commit/e3edc9746381773b5c57f17d1fdd8185d9ce3d11))
+* **DIST-701:** Load embed options from element attributes ([425e5c4](https://github.com/Typeform/embed/commit/425e5c4668c6e0bf0dcdaedb5590598b40412cd9))
+* **DIST-702:** Add support for opacity option ([3dbb4c6](https://github.com/Typeform/embed/commit/3dbb4c626042830b9d9a0b356ba47b80de01361e))
+* **DIST-703:** Add support for disableTracking option ([d26006c](https://github.com/Typeform/embed/commit/d26006c2d6b639c568eb44bca10d6dc3f495ef4f))
+* Add demo for react apps ([8dca211](https://github.com/Typeform/embed/commit/8dca2117345a50805034e9bb20ee8cea41a88536))
+* Initialize embed based on HTML attributes and auto-inject CSS ([47dffdf](https://github.com/Typeform/embed/commit/47dffdfb2a568c3b7b22fa539f2a7950be011f42))
+
+# [@typeform/embed-v1.0.0-next.2](https://github.com/Typeform/embed/compare/@typeform/embed-v1.0.0-next.1...@typeform/embed-v1.0.0-next.2) (2021-02-19)
+
+
+### Bug Fixes
+
+* **DIST-730:** Fix redraw issue ([a65076d](https://github.com/Typeform/embed/commit/a65076d057e003f1664f701633a43393d3b270cc))
+
+
+### Features
+
+* **DIST-701:** Add functional tests and update demo ([e3edc97](https://github.com/Typeform/embed/commit/e3edc9746381773b5c57f17d1fdd8185d9ce3d11))
+* **DIST-701:** Load embed options from element attributes ([425e5c4](https://github.com/Typeform/embed/commit/425e5c4668c6e0bf0dcdaedb5590598b40412cd9))
+* Add demo for react apps ([8dca211](https://github.com/Typeform/embed/commit/8dca2117345a50805034e9bb20ee8cea41a88536))
+* Initialize embed based on HTML attributes and auto-inject CSS ([47dffdf](https://github.com/Typeform/embed/commit/47dffdfb2a568c3b7b22fa539f2a7950be011f42))
+* **DIST-0:** Added config for cc velocity ([#170](https://github.com/Typeform/embed/issues/170)) ([02f7e71](https://github.com/Typeform/embed/commit/02f7e710b6184fe625f2f909a5abead9681acfb6))
+
 # [@typeform/embed-v1.0.0-next.2](https://github.com/Typeform/embed/compare/@typeform/embed-v1.0.0-next.1...@typeform/embed-v1.0.0-next.2) (2021-02-17)